python_proxy/requirements.txt
fastapi
httpx
uvicorn
python_dotenv